refactor: extract assignment data rendering into own component
- #120 #125
